+++ packages/core/src/helpers/injectedProvider.ts
@@ -1,5 +1,6 @@
 import detectEthereumProvider from '@metamask/detect-provider'
 import { providers } from 'ethers'
+import { isWebSocketProvider } from './isWebSocketProvider'
 
 const GET_METAMASK_LINK = 'https://metamask.io/download.html'
 
@@ -22,6 +23,8 @@ export async function getInjectedProvider(getPollingInterval: (chaindId: number)
 
   const provider = new providers.Web3Provider(injectedProvider, 'any')
   const chainId = await provider.send('eth_chainId', [])
-  provider.pollingInterval = getPollingInterval(chainId)
+  if (!isWebSocketProvider(provider)) {
+    provider.pollingInterval = getPollingInterval(chainId)
+  }
   return provider
 }