+++ packages/core/src/hooks/useCall.test.tsx
@@ -1,60 +1,34 @@
-import { MockProvider } from '@ethereum-waffle/provider'
 import { Contract } from 'ethers'
 import { useCall } from '..'
 import { expect } from 'chai'
 import {
-  renderWeb3Hook,
   deployMockToken,
   MOCK_TOKEN_INITIAL_BALANCE,
-  SECOND_TEST_CHAIN_ID,
   SECOND_MOCK_TOKEN_INITIAL_BALANCE,
   getResultPropery,
   renderDAppHook,
   setupTestingConfig,
-  mineBlock,
 } from '../testing'
-import { ChainId } from '../constants/chainId'
 import { BigNumber } from 'ethers'
 import { deployContract } from 'ethereum-waffle'
-import { BlockNumberContract, RevertContract, doublerContractABI } from '../constants'
+import { BlockNumberContract, reverterContractABI, doublerContractABI } from '../constants'
 import waitForExpect from 'wait-for-expect'
 
 describe('useCall', () => {
-  const mockProvider = new MockProvider()
-  const secondMockProvider = new MockProvider({ ganacheOptions: { _chainIdRpc: SECOND_TEST_CHAIN_ID } as any })
-  const [deployer] = mockProvider.getWallets()
-  const [secondDeployer] = secondMockProvider.getWallets()
-  let token: Contract
-  let secondToken: Contract
-  let chainId: number
-  let blockNumberContract: Contract
-  let secondBlockNumberContract: Contract
-  let revertContract: Contract
-
-  beforeEach(async () => {
-    chainId = (await mockProvider.getNetwork()).chainId
-    token = await deployMockToken(deployer)
-    secondToken = await deployMockToken(secondDeployer, SECOND_MOCK_TOKEN_INITIAL_BALANCE)
-    blockNumberContract = await deployContract(deployer, BlockNumberContract)
-    secondBlockNumberContract = await deployContract(deployer, BlockNumberContract)
-    revertContract = await deployContract(deployer, RevertContract)
-  })
-
-  for (const multicallVersion of [1] as const) {
+  for (const multicallVersion of [1, 2] as const) {
     describe(`Multicall v${multicallVersion}`, () => {
       it('initial test balance to be correct', async () => {
-        const { result, waitForCurrent } = await renderWeb3Hook(
+        const { config, network1 } = await setupTestingConfig({ multicallVersion })
+        const token = await deployMockToken(network1.deployer)
+        const { result, waitForCurrent } = await renderDAppHook(
           () =>
             useCall({
               contract: token,
               method: 'balanceOf',
-              args: [deployer.address],
+              args: [network1.deployer.address],
             }),
           {
-            readonlyMockProviders: {
-              [chainId]: mockProvider,
-            },
-            multicallVersion,
+            config,
           }
         )
         await waitForCurrent((val) => val !== undefined)
@@ -63,7 +37,10 @@ describe('useCall', () => {
       })
 
       it('returns error on revert', async () => {
-        const { result, waitForCurrent } = await renderWeb3Hook(
+        const { config, network1 } = await setupTestingConfig({ multicallVersion })
+        const revertContract = await deployContract(network1.deployer, reverterContractABI)
+
+        const { result, waitForCurrent } = await renderDAppHook(
           () =>
             useCall({
               contract: revertContract,
@@ -71,10 +48,7 @@ describe('useCall', () => {
               args: [],
             }),
           {
-            readonlyMockProviders: {
-              [chainId]: mockProvider,
-            },
-            multicallVersion,
+            config,
           }
         )
         await waitForCurrent((val) => val !== undefined)
@@ -83,12 +57,22 @@ describe('useCall', () => {
       })
 
       it('multichain calls return correct initial balances', async () => {
-        await testMultiChainUseCall(token, [deployer.address], ChainId.Localhost, MOCK_TOKEN_INITIAL_BALANCE)
+        const { config, network1, network2 } = await setupTestingConfig({ multicallVersion })
+        const token = await deployMockToken(network1.deployer)
+        const secondToken = await deployMockToken(network2.deployer, SECOND_MOCK_TOKEN_INITIAL_BALANCE)
+        await testMultiChainUseCall(
+          token,
+          [network1.deployer.address],
+          network1.chainId,
+          MOCK_TOKEN_INITIAL_BALANCE,
+          config
+        )
         await testMultiChainUseCall(
           secondToken,
-          [secondDeployer.address],
-          SECOND_TEST_CHAIN_ID,
-          SECOND_MOCK_TOKEN_INITIAL_BALANCE
+          [network2.deployer.address],
+          network2.chainId,
+          SECOND_MOCK_TOKEN_INITIAL_BALANCE,
+          config
         )
       })
 
@@ -96,9 +80,11 @@ describe('useCall', () => {
         contract: Contract,
         args: string[],
         chainId: number,
-        endValue: BigNumber
+        endValue: BigNumber,
+        // eslint-disable-next-line no-undef
+        config: Awaited<ReturnType<typeof setupTestingConfig>>['config']
       ) => {
-        const { result, waitForCurrent } = await renderWeb3Hook(
+        const { result, waitForCurrent } = await renderDAppHook(
           () =>
             useCall(
               {
@@ -109,11 +95,7 @@ describe('useCall', () => {
               { chainId }
             ),
           {
-            readonlyMockProviders: {
-              [ChainId.Localhost]: mockProvider,
-              [SECOND_TEST_CHAIN_ID]: secondMockProvider,
-            },
-            multicallVersion,
+            config,
           }
         )
         await waitForCurrent((val) => val !== undefined)
@@ -122,12 +104,16 @@ describe('useCall', () => {
       }
 
       it('Properly handles two calls', async () => {
-        const { result, waitForCurrent, mineBlock } = await renderWeb3Hook(
+        const { config, network1 } = await setupTestingConfig({ multicallVersion })
+        const token = await deployMockToken(network1.deployer)
+        const blockNumberContract = await deployContract(network1.deployer, BlockNumberContract)
+
+        const { result, waitForCurrent } = await renderDAppHook(
           () => {
             const balance = useCall({
               contract: token,
               method: 'balanceOf',
-              args: [deployer.address],
+              args: [network1.deployer.address],
             })
             const block = useCall({
               contract: blockNumberContract,
@@ -138,21 +124,18 @@ describe('useCall', () => {
             return { balance, block }
           },
           {
-            readonlyMockProviders: {
-              [chainId]: mockProvider,
-            },
-            multicallVersion,
+            config,
           }
         )
 
-        const blockNumber = await mockProvider.getBlockNumber()
+        const blockNumber = await network1.provider.getBlockNumber()
 
         await waitForCurrent(({ balance, block }) => !!(balance && block))
         expect(result.error).to.be.undefined
         expect(getResultPropery(result, 'balance')).to.eq(MOCK_TOKEN_INITIAL_BALANCE)
         expect(getResultPropery(result, 'block')).to.eq(blockNumber)
 
-        await mineBlock()
+        await network1.mineBlock()
 
         await waitForExpect(() => {
           expect(getResultPropery(result, 'balance')).to.eq(MOCK_TOKEN_INITIAL_BALANCE)
@@ -161,7 +144,11 @@ describe('useCall', () => {
       })
 
       it('Properly handles refresh per block', async () => {
-        const { result, waitForCurrent, mineBlock } = await renderWeb3Hook(
+        const { config, network1 } = await setupTestingConfig({ multicallVersion })
+        const blockNumberContract = await deployContract(network1.deployer, BlockNumberContract)
+        const secondBlockNumberContract = await deployContract(network1.deployer, BlockNumberContract)
+
+        const { result, waitForCurrent } = await renderDAppHook(
           () => {
             const block1 = useCall({
               contract: blockNumberContract,
@@ -183,27 +170,24 @@ describe('useCall', () => {
             return { block1, block2 }
           },
           {
-            readonlyMockProviders: {
-              [chainId]: mockProvider,
-            },
-            multicallVersion,
+            config,
           }
         )
 
-        const blockNumber = await mockProvider.getBlockNumber()
+        const blockNumber = await network1.provider.getBlockNumber()
 
         await waitForCurrent(({ block1, block2 }) => !!(block1 && block2))
         expect(result.error).to.be.undefined
         expect(getResultPropery(result, 'block1')).to.eq(blockNumber)
         expect(getResultPropery(result, 'block2')).to.eq(blockNumber)
 
-        await mineBlock()
+        await network1.mineBlock()
 
         await waitForCurrent(({ block1 }) => block1 !== undefined && block1.value[0].toNumber() === blockNumber + 1)
         expect(getResultPropery(result, 'block1')).to.eq(blockNumber + 1)
         expect(getResultPropery(result, 'block2')).to.eq(blockNumber)
 
-        await mineBlock()
+        await network1.mineBlock()
 
         await waitForExpect(() => {
           expect(getResultPropery(result, 'block1')).to.eq(blockNumber + 2)
@@ -211,7 +195,7 @@ describe('useCall', () => {
         })
 
         for (let i = 0; i < 3; i++) {
-          await mineBlock()
+          await network1.mineBlock()
         }
 
         await waitForExpect(() => {
@@ -223,9 +207,8 @@ describe('useCall', () => {
       })
 
       it('Refreshes static call on parameter change', async () => {
-        const { config, network1 } = await setupTestingConfig()
-        const [deployer] = network1.provider.getWallets()
-        const doublerContract = await deployContract(deployer, doublerContractABI)
+        const { config, network1 } = await setupTestingConfig({ multicallVersion })
+        const doublerContract = await deployContract(network1.deployer, doublerContractABI)
         const { waitForCurrent, rerender } = await renderDAppHook(
           ({ num }: { num: number }) =>
             useCall({
@@ -255,9 +238,8 @@ describe('useCall', () => {
 
       it('Refreshes only static calls with changed parameter', async () => {
         const { config, network1 } = await setupTestingConfig()
-        const [deployer] = network1.provider.getWallets()
-        const doublerContract = await deployContract(deployer, doublerContractABI)
-        const blockNumberContract = await deployContract(deployer, BlockNumberContract)
+        const doublerContract = await deployContract(network1.deployer, doublerContractABI)
+        const blockNumberContract = await deployContract(network1.deployer, BlockNumberContract)
         const { waitForCurrent, rerender, result } = await renderDAppHook(
           ({ num }: { num: number }) => {
             const doubled = useCall({
@@ -290,7 +272,8 @@ describe('useCall', () => {
         await waitForCurrent((val) => val?.doubled?.value?.[0]?.eq(2))
         const blockNumberBefore = result.current.blockNumber?.value[0]
 
-        await mineBlock(network1.provider)
+        await network1.mineBlock()
+
         expect(result.current.doubled?.value[0]).to.eq(2)
         expect(result.current.blockNumber?.value[0]).to.eq(blockNumberBefore)
         rerender({ num: 2 })